IAFW and SECFW versions in Android properties
BZ: 136275

Data in /sys/firmware/pidv/iafw_version and
/sys/firmware/pidv/secfw_version do not match with values
set by IA Firmware.

Reviewed-on: http://android.intel.com:8080/130129
Tested-by: Lebouc, Christophe <christophe.lebouc@intel.com>
Tested-by: cactus <cactus@intel.com>

+++ include/linux/intel_mid_acpi.h
@@ -41,8 +41,8 @@ struct platform_id {
 	u8 ext_id_1[32];
 	u8 ext_id_2[32];
 	u8 uuid[16];
-	u8 iafw_major;
-	u8 iafw_minor;
+	u16 iafw_major;
+	u16 iafw_minor;
 	u8 secfw_major;
 	u8 secfw_minor;
 };
